Change __callapi function
Most calls are POST requests, so add that branch of logic to function,
make default to it doesn’t need to be specified in functions.

Every function stub SHOULD be a POST request, but verify. For easier
coding, I added a debug argument to the Class constructor, which will
print out the body of the function (which should match what the API
explorer expects)